Set TotalPages to 1 when error occurs
If an error occurs, we need to set the TotalPages to 1 by default and use it to calculate the allotted timeout. The reason why we need to
set it to 1 because we don't want Hylafax to decide whether or not the fax fails immediately, so we set the pages count to 1 so it at least has a chance of sending that means the timeout will be set to 400. If the pages count was read as 0 then it's likely an error, or it wasn't set for some reason.


Signed-off-by: michael <michael.suarez@synacy.com>